requirements.txt
bs4
requests_html
lxml